Move validation and handler to page level
